Additionally provide typesafe setters for kotlin-dsl
Right now, when using SpongeGradle with kotlin-dsl, it was not
possible to use Kotlin's standard property assignment syntax.
As an example, instead of
    plugin.meta { name = "MyPlugin" }
you had to do
    plugin.meta { setName("MyPlugin") }

This is because Kotlin detects properties based on a getter and
setter of the same type but these are defined using a String
getter and a generic Object setter (e.g. to allow passing closures).

However, to make Kotlin's property assignment syntax work we can
just define setters for both types - Object and String - and have
nice syntax for both Groovy and Kotlin. Yay!