SpongeGradle 0.9.0
There were some problems with an old version of the Gradle plugin-publish
plugin, causing 0.8.2 to be uploaded in a broken state on the Gradle
plugin portal. It was fixed by deleting the old version and re-uploading
using a newer version of the plugin-publish plugin but it seems like
Gradle still serves the broken version for some users (probably due to
caching).

Given that there were new additions in the last release, 0.9.0 is a more
appropriate version number anyway. So re-release 0.8.2 as 0.9.0 to make
it work for everyone. :)