Redefine all extension variables to recommended Gradle
Previously, all fields were configured immediately, but
with Gradle 4 and newer, we're being driven towards a lazy
configuration paradigm. This means that we have to defer
everything that previously we expected to be available to
be configured "as-needed". This affects how the versions
are compiled and how tasks are configured based on api
and common projects. What this also does for us is allow
for our `common` and `impl` targets to loosely define where
a sub project exists, let alone how the project ends up
being defined folder structure wise.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>