Lazily configure even more sections of the plugin.
Cleaning up the SourceType to better handle sourceSets
as they are being added, the idea with them being lazily
configured results in some time insensitive configuration
for all sources and source sets. As it stands currently,
nested project dependencies are being fubarred a little,
and Invalid still doesn't function the way it should.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>